Implements TableOne classes
TableOne will be moved to its own library as it's going to be useful in
other projects as well. It implements tables in the command line output
(but could be used for comment generation or documentation generation as
well, and maybe extended for other kinds of output).

It supports "styling" and the plan is in the future to support colors
and other advanced formatting of tables, such as multi-row and
multi-column cells.